Opinion Event Selector
This is a project targets to help researchers and public opinion analysts to evaluate and find the best event mix for study. The benchmark and criteria for a satisfactory event mix are as follows:
1. The events covers as many fields, risk types, and subject entity categories as possible
2. The popularity of each event is evenly distributed, in case that some extremely viral events would affect the outcomes of opinion study, and leading to a miscomprehension of opinion distribution and extremity
3. In order to avoid the overlap in events and opinion leaders, the event mix tend to cover different events that are inherently unrelated. To be specific, if event A and B are sub-events of a big event, then only one of them shall be selected into the mix
4. Based on the selection criteria, the emotionality selector and filter can be added to optimize the process

Research and Project Target:
Providing an automate tool to group social events for sociology, communication, information management, and corporate management studies. The selected events shall be representative to reflect how the publics respond to evaluates a certain brand, or how opinion are formed and developed on a certain opinion market (such as X(Twitter), Instagram, Weibo, or WeChat).

However, due to my limited NLP ability, more work could be done, and I am welcomed to any revisions and suggestions on the project.

Potential algorithmic optimization directions:
1. how to evaluate the product mix and design the evaluator?
2. How to better identify the closeness and familiarities between events, and how to check and test events?